+++ rtv/objects.py
@@ -62,7 +62,8 @@ def curses_session():
         # Hide the blinking cursor
         curses.curs_set(0)
 
-        Color.init()
+        # Assign the terminal's default (background) color to code -1
+        curses.use_default_colors()
 
         yield stdscr
 
@@ -254,7 +255,6 @@ class LoadScreen(object):
 
 
 class Color(object):
-
     """
     Color attributes for curses.
     """
@@ -286,9 +286,6 @@ class Color(object):
         curses color pairs can be accessed directly through class attributes.
         """
 
-        # Assign the terminal's default (background) color to code -1
-        curses.use_default_colors()
-
         for index, (attr, code) in enumerate(cls._colors.items(), start=1):
             curses.init_pair(index, code[0], code[1])
             setattr(cls, attr, curses.color_pair(index))
@@ -318,7 +315,8 @@ class Navigator(object):
             valid_page_cb,
             page_index=0,
             cursor_index=0,
-            inverted=False):
+            inverted=False,
+            top_item_height=None):
         """
         Params:
             valid_page_callback (func): This function, usually `Content.get`,
@@ -334,11 +332,16 @@ class Navigator(object):
                 inverted - The page is drawn from the bottom of the screen,
                     starting with the page index, up to the top of the
                     screen.
+            top_item_height (int): If this is set to a non-null value
+            The number of columns that the top-most item
+                should utilize if non-inverted. This is used for a special mode
+                where all items are drawn non-inverted except for the top one.
         """
 
         self.page_index = page_index
         self.cursor_index = cursor_index
         self.inverted = inverted
+        self.top_item_height = top_item_height
         self._page_cb = valid_page_cb
 
     @property
@@ -399,15 +402,21 @@ class Navigator(object):
                     # Flip the orientation and reset the cursor
                     self.flip(self.cursor_index)
                     self.cursor_index = 0
+                    self.top_item_height = None
                     redraw = True
         else:
             if self.cursor_index > 0:
                 self.cursor_index -= 1
+                if self.top_item_height and self.cursor_index == 0:
+                    # Selecting the partially displayed item
+                    self.top_item_height = None
+                    redraw = True
             else:
                 self.page_index -= self.step
                 if self._is_valid(self.absolute_index):
                     # We have reached the beginning of the page - move the
                     # index
+                    self.top_item_height = None
                     redraw = True
                 else:
                     self.page_index += self.step
@@ -481,6 +490,7 @@ class Navigator(object):
         self.page_index += (self.step * n_windows)
         self.cursor_index = n_windows
         self.inverted = not self.inverted
+        self.top_item_height = None
 
     def _is_valid(self, page_index):
         """