+++ CHANGELOG.rst
@@ -2,6 +2,9 @@
 RTV Changelog
 =============
 
+.. _1.10.0: http://github.com/michael-lazar/rtv/releases/tag/v1.10.0
+.. _1.9.1: http://github.com/michael-lazar/rtv/releases/tag/v1.9.1
+.. _1.9.0: http://github.com/michael-lazar/rtv/releases/tag/v1.9.0
 .. _1.8.1: http://github.com/michael-lazar/rtv/releases/tag/v1.8.1
 .. _1.8.0: http://github.com/michael-lazar/rtv/releases/tag/v1.8.0
 .. _1.7.0: http://github.com/michael-lazar/rtv/releases/tag/v1.7.0
@@ -15,6 +18,56 @@ RTV Changelog
 .. _1.2.2: http://github.com/michael-lazar/rtv/releases/tag/v1.2.2
 .. _1.2.1: http://github.com/michael-lazar/rtv/releases/tag/v1.2.1
 .. _1.2: http://github.com/michael-lazar/rtv/releases/tag/v1.2
+
+--------------------
+1.10.0_ (2016-07-11)
+--------------------
+Features
+
+* New command, `b` extracts urls from comments using urlviewer.
+* Comment files will no longer be destroyed if RTV encounters an error while posting.
+* The terminal title now displays the subreddit name/url.
+
+Bugfixes
+
+* Fixed crash when entering empty or invalid subreddit name.
+* Fixed crash when opening x-posts linked to subreddits.
+* Fixed a bug where the terminal title wasn't getting set.
+* **/r/me** is now displayed as *My Submissions* in the header.
+
+-------------------
+1.9.1_ (2016-06-13)
+-------------------
+Features
+
+* Better support for */r/random*.
+* Added a ``monochrome`` config setting to disable all color.
+* Improved cursor positioning when expanding/hiding comments.
+* Show ``(not enough space)`` when comments are too large.
+
+Bugfixes
+
+* Fixed permissions when copying the config file.
+* Fixed bug where submission indicies were duplicated when paging.
+* Specify praw v3.4.0 to avoid installing praw 4.
+
+Documentation
+
+* Added section to the readme on Arch Linux installation.
+* Updated a few argument descriptions.
+* Added a proper ascii logo.
+
+-------------------
+1.9.0_ (2016-04-05)
+-------------------
+Features
+
+* You can now open long posts/comments with the $PAGER by pressing `l`.
+* Changed a couple of visual separators.
+
+Documentation
+
+* Added testing instructions to the FAQ.
 
 -------------------
 1.8.1_ (2016-03-01)